backend/requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
python-dotenv
prometheus_client
